Refactor JavaScript test suite and related functionality
- Removed the old js.cc file and its associated logic.
- Introduced new files for organizing JavaScript test cases:
  - auto.cc: Handles automatic JavaScript test case execution.
  - file.cc: Contains logic for running service JavaScript callbacks.
  - http.cc: Implements an HTTP server for testing JavaScript HTTP client functionality.
  - js_case.cc: Centralizes JavaScript test case utilities and file loading.
- Updated js_case.h to declare new utility functions for loading and running JavaScript test cases.
- Added type.cc to implement API tests for various JavaScript types and conversions.
- Improved error handling and test case registration in the new structure.
